outlook_mailer/pyproject.toml
[project]
name = "outlook-mailer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "pytz>=2024.1",
]
